Initial commit of 0.9.0 code
This is the almost-exact copy of our current code in production on
mysteryscience.com. The Sidekiq specifics were altered in a blunt way
to bypass all of its code since we can't have that as a dependency in
the core gem. We'll need to figure out a plugin strategy, whether
that's separate gems (probably too soon for all that) or some sort of
runtime checks for the presence of 3rd party libraries.

There is a Slack specific class, but it's been designed to require a
subclass that will handle the actual integration to Slack. That's not a
great experience either.

The goal for a 1.0 release of the gem would be to have a suitable plugin
strategy.